+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Add
 
 * Users listing
+* Collections listing
 
 
 ## [0.3.7] - 2019-06-14